requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
google-cloud-storage==1.42.0
google-cloud-bigquery==2.32.0
google-cloud-pubsub==2.13.12
google-api-python-client==2.37.0
psycopg2-binary==2.9.1
google-auth==1.31.0
requests==2.26.0
pyyaml==6.0